% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{reliability_calculator}
\alias{reliability_calculator}
\title{Map reliability calculator}
\usage{
reliability_calculator(
  data,
  est,
  moe,
  viewer = getOption(x = "mapreliability.viewer", default = "pane")
)
}
\arguments{
\item{data}{Dataframe}

\item{est}{Estimate column}

\item{moe}{Margin of error column}

\item{viewer}{Where to display the gadget: \code{"dialog"}, \code{"pane"} or
\code{"browser"} (see \code{\link[shiny:viewer]{shiny::viewer()}}).}
}
\description{
Takes a dataframe with estimate and MOE columns and launches an RStudio addin
to explore the reliability of various classifications of the estimate values for mapping.
}
\examples{
\dontrun{

if (interactive()) {

bk_hh_income <- tidycensus::get_acs(
  geography = "tract",
  state = "NY",
  county = "Kings",
  year = 2018,
  survey = "acs5",
  variables = "B19013_001", # Median Household Income
  key = Sys.getenv("CENSUS_API")
)

reliability_calculator(bk_hh_income, estimate, moe)

}

}
}
