% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability_table}
\alias{reliability_table}
\alias{reliability_table_equal}
\alias{reliability_table_quant}
\alias{reliability_table_custom}
\title{Classification reliability tables}
\usage{
reliability_table_equal(
  data,
  est,
  moe,
  n_classes,
  conf_level = 0.9,
  quiet = FALSE
)

reliability_table_quant(
  data,
  est,
  moe,
  n_classes,
  conf_level = 0.9,
  quiet = FALSE
)

reliability_table_custom(
  data,
  est,
  moe,
  class_breaks_low,
  conf_level = 0.9,
  quiet = FALSE
)
}
\arguments{
\item{data}{Dataframe}

\item{est}{Estimate column}

\item{moe}{Margin of error column}

\item{n_classes}{Number of classes}

\item{conf_level}{Confidence level used for MOEs (default = 90\\%)}

\item{quiet}{Whether to raise warnings about dropped observations with
missing data}

\item{class_breaks_low}{A vector of lower break points for a custom classification}
}
\description{
Get a table of reliability results for a classification of a set of
estimates.
}
\examples{
\dontrun{
bk_hh_income <- tidycensus::get_acs(
  geography = "tract",
  state = "NY",
  county = "Kings",
  year = 2018,
  survey = "acs5",
  variables = "B19013_001", # Median Household Income
  key = Sys.getenv("CENSUS_API")
)

reliability_table_equal(bk_hh_income, estimate, moe, 4)

reliability_table_quant(bk_hh_income, estimate, moe, 4)

reliability_table_custom(bk_hh_income, estimate, moe, c(0, 40000, 80000))

}

}
